rename package from opentelemetry to tracing
