+++ modules/caddyhttp/tracing/tracerprovider.go
@@ -1,4 +1,4 @@
-package opentelemetry
+package tracing
 
 import (
 	"context"
